foo
Former-commit-id: 5378fc807aa58ae5d69b1df667fc0539ae789962 [formerly 48d2bbb9b374ba37d01f9a78cb792983ed9f5d1e]
Former-commit-id: f8d6366de0dc7a42d97595c33aa9bbdde223ee5d

+++ twitter/topUsers/createTables.sql
@@ -9,7 +9,7 @@ create table if not exists topUsers (
 drop table if exists tweets cascade;
 create table if not exists tweets (
 	id bigint not null,
-	user_id integer not null,
+	user_id bigint not null,
 	user_screen_name varchar(255) not null,
 	created_at varchar(255) not null,
 	retweeted varchar(255) not null,
